+++ shared/src/nodes/func_ty.rs
@@ -291,6 +291,23 @@ pub struct Function {
     pub body: Option<Statement>,
     pub params: ParameterList,
     pub returns: ParameterList,
+    pub modifiers_set: bool,
+}
+
+impl Default for Function {
+    fn default() -> Self {
+        Self {
+            loc: Loc::Implicit,
+            ty: FunctionTy::Function,
+            name: None,
+            name_loc: Loc::Implicit,
+            attributes: vec![],
+            body: None,
+            params: vec![],
+            returns: vec![],
+            modifiers_set: true,
+        }
+    }
 }
 
 impl Function {
@@ -304,43 +321,32 @@ impl Function {
         if modifiers.is_empty() {
             vec![]
         } else {
-            modifiers
-                .iter()
-                .filter_map(|modifier| {
-                    assert_eq!(modifier.name.identifiers.len(), 1);
-                    // construct arg string for function selector
-                    let mut mod_name = format!("{}(", modifier.name.identifiers[0]);
-                    if let Some(args) = &modifier.args {
-                        let args_str = args
-                            .iter()
-                            .map(|expr| match expr {
-                                Expression::Variable(ident) => {
-                                    let func_param_ty_str = func_params
-                                        .iter()
-                                        .find(|param| {
-                                            if let Some(param_name) = param.maybe_name(analyzer) {
-                                                param_name == ident.name
-                                            } else {
-                                                false
-                                            }
-                                        })
-                                        .map(|param| param.ty_str(analyzer))
-                                        .unwrap_or_else(|| "".to_string());
-                                    func_param_ty_str
-                                }
-                                e => todo!("here: {e:?}"),
-                            })
-                            .collect::<Vec<_>>()
-                            .join(", ");
-                        let _ = write!(mod_name, "{}", args_str);
-                    }
-                    let _ = write!(mod_name, ")");
-                    analyzer
-                        .user_types()
-                        .get(&mod_name)
-                        .map(|idx| FunctionNode::from(*idx))
-                })
-                .collect()
+            modifiers.iter().filter_map(|modifier| {
+                assert_eq!(modifier.name.identifiers.len(), 1);
+                // construct arg string for function selector
+                let mut mod_name = format!("{}(", modifier.name.identifiers[0]);
+                if let Some(args) = &modifier.args {
+                    let args_str = args.iter().map(|expr| {
+                        match expr {
+                            Expression::Variable(ident) => {
+                                let func_param_ty_str = func_params.iter().find(|param| {
+                                    if let Some(param_name) = param.maybe_name(analyzer) {
+                                        param_name == ident.name
+                                    } else {
+                                        false
+                                    }
+                                }).map(|param| param.ty_str(analyzer)).unwrap_or_else(|| "".to_string());
+                                func_param_ty_str
+                            }
+                            // Expression::FunctionCall()
+                            e => todo!("here: {e:?}")
+                        }
+                    }).collect::<Vec<_>>().join(", ");
+                    let _ = write!(mod_name, "{}", args_str);
+                }
+                let _ = write!(mod_name, ")");
+                analyzer.user_types().get(&mod_name).map(|idx| FunctionNode::from(*idx))
+            }).collect()
         }
     }
 
@@ -372,6 +378,7 @@ impl From<FunctionDefinition> for Function {
             body: func.body,
             params: func.params,
             returns: func.returns,
+            modifiers_set: false,
         }
     }
 }
@@ -397,15 +404,8 @@ impl From<VariableDefinition> for Function {
                 )],
             }),
             params: vec![],
-            returns: vec![(
-                var.loc,
-                Some(Parameter {
-                    loc: var.loc,
-                    ty: var.ty,
-                    storage: None,
-                    name: None,
-                }),
-            )],
+            returns: vec![(var.loc, Some(Parameter { loc: var.loc, ty: var.ty, storage: None, name: None }))],
+            modifiers_set: true,
         }
     }
 }
